requirements.txt
openai
sounddevice
numpy
pyttsx3
python-dotenv
pypiwin32; sys_platform == 'win32'
pyobjc; sys_platform == 'darwin'
